add: create_model sub-package
